Seeders
-Se crean los seeder para el llenado de cada tabla correspondientes a
los modelos, con datos falsos a travez de "Faker".

+++ database/migrations/2016_11_04_171504_posicion_migration.php
@@ -14,8 +14,8 @@ class PosicionMigration extends Migration
     {
         Schema::create('posiciones', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('lat');
-            $table->integer('lan');
+            $table->double('lat');
+            $table->double('lan');
             $table->dateTime('fecha_registro');
             $table->integer('tracker_id')->unsigned();
             $table->foreign('tracker_id')->references('id')->on('trackers');
